pyproject.toml
[project]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = [
    {name = "Pavel",email = "shegol164@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pytest (>=8.3.5,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.14.1"
black = "^25.1.0"
isort = "^6.0.0"

